controls namespace refactoring
