+++ MediaCenter/MediaCenter/MainWindow.xaml
@@ -4,11 +4,10 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:MediaCenter"
-        xmlns:sessions="clr-namespace:MediaCenter.Sessions"
         xmlns:staging="clr-namespace:MediaCenter.Sessions.Staging"
         xmlns:query="clr-namespace:MediaCenter.Sessions.Query"
         xmlns:repository="clr-namespace:MediaCenter.Repository"
-        xmlns:controls="clr-namespace:MediaCenter.Controls"
+        xmlns:controls="clr-namespace:MediaCenter.WPF.Controls"
         mc:Ignorable="d"
         d:DataContext="{d:DesignInstance local:MainWindowViewModel}"
         Title="{Binding Path=AppVersion, StringFormat={}Media Center - v {0}}" Height="350" Width="525" WindowState="Maximized" Background="DarkSeaGreen">
