+++ MediaCenter/MediaCenter/Media/ImageItemContentView.xaml
@@ -4,7 +4,7 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:local="clr-namespace:MediaCenter.Media"
-             xmlns:controls="clr-namespace:MediaCenter.Controls"
+             xmlns:controls="clr-namespace:MediaCenter.WPF.Controls"
              mc:Ignorable="d" 
              d:DesignHeight="300" d:DesignWidth="300"
              d:DataContext="{d:DesignInstance Type=local:ImageItemViewModel}">
