+++ MediaCenter/MediaCenter/Sessions/Filters/FilterCollectionView.xaml
@@ -4,8 +4,7 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:local="clr-namespace:MediaCenter.Sessions.Filters"
-             xmlns:controls="clr-namespace:MediaCenter.Controls"
-             xmlns:styles="clr-namespace:MediaCenter.Styles"
+             xmlns:controls="clr-namespace:MediaCenter.WPF.Controls"
              mc:Ignorable="d" 
              d:DesignHeight="300" d:DesignWidth="300"
              d:DataContext="{d:DesignInstance local:FilterCollectionViewModel}">
