+++ MediaCenter/MediaCenter/Sessions/Staging/StagingSessionView.xaml
@@ -5,14 +5,14 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:i="http://schemas.microsoft.com/expression/2010/interactivity"
              xmlns:local="clr-namespace:MediaCenter.Sessions.Staging"
-             xmlns:styles="clr-namespace:MediaCenter.Styles"
              xmlns:media="clr-namespace:MediaCenter.Media"
+             xmlns:wpf="clr-namespace:MediaCenter.WPF"
              mc:Ignorable="d" 
              d:DataContext="{d:DesignInstance local:StagingSessionViewModel}"
              d:DesignHeight="300" d:DesignWidth="300">
     
     <i:Interaction.Behaviors>
-        <styles:OpenCloseWindowBehavior Trigger="{Binding ShowEditViewModel}" WindowType="local:EditStagedItemView" WindowDataContext="{Binding EditViewModel}" 
+        <wpf:OpenCloseWindowBehavior Trigger="{Binding ShowEditViewModel}" WindowType="local:EditStagedItemView" WindowDataContext="{Binding EditViewModel}" 
                                         OwnerWindow="{Binding RelativeSource={RelativeSource Mode=FindAncestor, AncestorType=Window}}"/>
     </i:Interaction.Behaviors>
