+++ MediaCenter/MediaCenter/WPF/Controls/CustomControls.xaml
@@ -1,6 +1,6 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:local="clr-namespace:MediaCenter.Controls">
+                    xmlns:local="clr-namespace:MediaCenter.WPF.Controls">
     <ResourceDictionary.MergedDictionaries>
         <ResourceDictionary Source="ImageButton.xaml"/>
         <ResourceDictionary Source="ImageToggleButton.xaml"/>
