+++ MediaCenter/MediaCenter/WPF/Controls/ImageButton.cs
@@ -1,13 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Media;
 
-namespace MediaCenter.Controls
+namespace MediaCenter.WPF.Controls
 {
     public class ImageButton : Button
     {static ImageButton()
