+++ MediaCenter/MediaCenter/WPF/Controls/VideoPlayer.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="MediaCenter.Controls.VideoPlayer"
+﻿<UserControl x:Class="MediaCenter.WPF.Controls.VideoPlayer"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
