+++ MediaCenter/MediaCenter/WPF/Controls/ZoomImage.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="MediaCenter.Controls.ZoomImage"
+﻿<UserControl x:Class="MediaCenter.WPF.Controls.ZoomImage"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:MediaCenter.Controls"
+             xmlns:local="clr-namespace:MediaCenter.WPF.Controls"
              mc:Ignorable="d" 
              d:DesignHeight="300" d:DesignWidth="300">
     <Grid>
