+++ MediaCenter/MediaCenter/WPF/ScrollIntoViewBehavior.cs
@@ -1,7 +1,7 @@
 ﻿using System.Windows.Controls;
 using System.Windows.Interactivity;
 
-namespace MediaCenter.Styles
+namespace MediaCenter.WPF
 {
     public sealed class ScrollIntoViewBehavior:Behavior<ListBox>
     {
